requirements.txt
langchain
langchain-openai
langchain_community
pillow
streamlit
dotenv
pydantic
beautifulsoup4
python-slugify